+++ crates/store-dir/src/cas_file.rs
@@ -5,6 +5,15 @@ use pacquet_fs::{ensure_file, make_file_executable, EnsureFileError};
 use sha2::{Digest, Sha512};
 use std::path::PathBuf;
 
+impl StoreDir {
+    /// Path to a file in the store directory.
+    pub fn cas_file_path(&self, hash: FileHash, executable: bool) -> PathBuf {
+        let hex = format!("{hash:x}");
+        let suffix = if executable { "-exec" } else { "" };
+        self.file_path_by_hex_str(&hex, suffix)
+    }
+}
+
 /// Error type of [`StoreDir::write_cas_file`].
 #[derive(Debug, Display, Error, Diagnostic)]
 pub enum WriteCasFileError {
@@ -25,5 +34,35 @@ impl StoreDir {
         make_file_executable(&file_path).expect("make the file executable"); // TODO: propagate this error
 
         Ok((file_path, file_hash))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn cas_file_path() {
+        fn case(file_content: &str, executable: bool, expected: &str) {
+            eprintln!("CASE: {file_content:?}, {executable:?}");
+            let store_dir = StoreDir::new("STORE_DIR");
+            let file_hash = Sha512::digest(file_content);
+            eprintln!("file_hash = {file_hash:x}");
+            let received = store_dir.cas_file_path(file_hash, executable);
+            let expected: PathBuf = expected.split('/').collect();
+            assert_eq!(&received, &expected);
+        }
+
+        case(
+            "hello world",
+            false,
+            "STORE_DIR/v3/files/30/9ecc489c12d6eb4cc40f50c902f2b4d0ed77ee511a7c7a9bcd3ca86d4cd86f989dd35bc5ff499670da34255b45b0cfd830e81f605dcf7dc5542e93ae9cd76f",
+        );
+
+        case(
+            "hello world",
+            true,
+            "STORE_DIR/v3/files/30/9ecc489c12d6eb4cc40f50c902f2b4d0ed77ee511a7c7a9bcd3ca86d4cd86f989dd35bc5ff499670da34255b45b0cfd830e81f605dcf7dc5542e93ae9cd76f-exec",
+        );
     }
 }
